version 0.6

+++ src/main/java/com/crocostaud/stockmanagement/controller/ShopController.java
@@ -4,8 +4,8 @@ import com.crocostaud.stockmanagement.dto.stock.ShopDto;
 import com.crocostaud.stockmanagement.model.stock.ShopUser;
 import com.crocostaud.stockmanagement.service.ShopService;
 import com.crocostaud.stockmanagement.service.UserService;
-import com.crocostaud.stockmanagement.utils.annotation.Username;
-import com.crocostaud.stockmanagement.utils.security.Auth;
+import com.crocostaud.stockmanagement.utils.annotation.User;
+import com.crocostaud.stockmanagement.utils.request.ShopRequest;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -17,32 +17,26 @@ import org.springframework.web.server.ResponseStatusException;
 public class ShopController {
     private final ShopService shopService;
     private final UserService userService;
-    private final Auth auth;
 
-    public ShopController(ShopService shopService, Auth auth, UserService userService) {
+    public ShopController(ShopService shopService, UserService userService) {
         this.shopService = shopService;
-        this.auth = auth;
         this.userService = userService;
     }
 
     @PostMapping
-    public ResponseEntity<ShopDto> create(@RequestBody ShopDto shopDto, @Username String username) {
-        ShopUser user = auth.getUser(username);
+    public ResponseEntity<ShopDto> create(@RequestBody ShopRequest shopRequest, @User ShopUser user) {
         if (user == null || user.getShop() != null) {
             return ResponseEntity.badRequest().build();
         }
-        ShopDto savedShop = shopService.createShop(shopDto);
+        ShopDto savedShop = shopService.createShop(shopRequest.shopDto(), shopRequest.warehouseDto());
         userService.setShop(user.getId(), savedShop.getId());
         return ResponseEntity.ok(savedShop);
     }
 
     @PutMapping
-    public ResponseEntity<ShopDto> update(@RequestBody ShopDto shopDto, @Username String username) {
-        ShopUser user = auth.getUser(username);
-
+    public ResponseEntity<ShopDto> update(@RequestBody ShopDto shopDto, @User ShopUser user) {
         if (user == null || user.getShop() == null || user.getShop().getId() == null)
             throw new ResponseStatusException(HttpStatusCode.valueOf(401), "User not found");
-
 
         if (!user.getShop().getId().equals(shopDto.getId()))
             throw new ResponseStatusException(HttpStatusCode.valueOf(401), "Shop not found");
@@ -52,11 +46,8 @@ public class ShopController {
     }
 
     @GetMapping
-    public ResponseEntity<ShopDto> get(@Username String username) {
-        ShopUser user = auth.getUser(username);
-        System.out.println(username + "  -> " + user);
-
-        ShopDto shop = shopService.getShop(username);
+    public ResponseEntity<ShopDto> get(@User ShopUser user) {
+        ShopDto shop = shopService.getShop(user.getUsername());
 
         if (shop == null)
             return ResponseEntity.noContent().build();
@@ -65,14 +56,11 @@ public class ShopController {
     }
 
     @DeleteMapping("/{shopId}")
-    public ResponseEntity<ShopDto> delete(@Username String username, @PathVariable Long shopId) {
-        ShopUser user = auth.getUser(username);
-//        if (user == null || user.getShop() == null) {
-//            return ResponseEntity.status(HttpStatusCode.valueOf(400)).build();
-//        }
+    public ResponseEntity<ShopDto> delete(@PathVariable Long shopId, @User ShopUser user) {
+        if (user == null || user.getShop() == null) {
+            return ResponseEntity.status(HttpStatusCode.valueOf(400)).build();
+        }
         ShopDto shop = shopService.getShop(shopId);
         return ResponseEntity.ok(shopService.delete(shop.getId()));
-
     }
-
 }
